Extract non-DEM preserved material grouping
