+++ src/PlateauResoniteLink/Targets/Resonite/NonDemCityObjectBaker.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -495,10 +494,10 @@ internal sealed class NonDemCityObjectBaker(
                     CommonMaterial: CommonMaterialCatalog.Create().Generic.Uv));
         }
 
-        foreach (IGrouping<PreservedMaterialGroupingKey, OrderedPreservedSubmeshEntry> preservedGroup in candidates
+        foreach (IGrouping<NonDemPreservedMaterialGroupingKey, OrderedPreservedSubmeshEntry> preservedGroup in candidates
                      .SelectMany(static candidate => candidate.PreservedEntries)
                      .Select(static (entry, order) => new OrderedPreservedSubmeshEntry(entry, order))
-                     .GroupBy(static entry => CreatePreservedMaterialGroupingKey(entry.Entry.Material), PreservedMaterialGroupingKeyComparer.Instance)
+                     .GroupBy(static entry => NonDemPreservedMaterialGrouping.CreateKey(entry.Entry.Material), NonDemPreservedMaterialGrouping.KeyComparer)
                      .OrderBy(static group => group.Min(static entry => entry.Order)))
         {
             cancellationToken.ThrowIfCancellationRequested();
@@ -517,7 +516,7 @@ internal sealed class NonDemCityObjectBaker(
             }
 
             int submeshIndex = submeshes.Count;
-            ResoniteMaterialBinding preservedMaterial = NormalizePreservedMaterial(preservedGroup.First().Entry.Material) with
+            ResoniteMaterialBinding preservedMaterial = NonDemPreservedMaterialGrouping.NormalizeMaterial(preservedGroup.First().Entry.Material) with
             {
                 SubmeshIndices = [submeshIndex],
             };
@@ -1161,50 +1160,6 @@ internal sealed class NonDemCityObjectBaker(
         return (byte)Math.Clamp(Math.Round(value), 0.0, 255.0);
     }
 
-    private static PreservedMaterialGroupingKey CreatePreservedMaterialGroupingKey(ResoniteMaterialBinding material)
-    {
-        ResoniteMaterialBinding normalizedMaterial = NormalizePreservedMaterial(material);
-        if (normalizedMaterial.CommonMaterial is not null)
-        {
-            return new PreservedMaterialGroupingKey(
-                normalizedMaterial.CommonMaterial,
-                new ResoniteColor(0.0, 0.0, 0.0, 0.0),
-                default,
-                normalizedMaterial.TexturePayload,
-                normalizedMaterial.TextureSourceKind,
-                normalizedMaterial.TerrainOverlay,
-                default,
-                default,
-                default,
-                default,
-                default,
-                default,
-                default,
-                normalizedMaterial.TerrainMeshCode);
-        }
-
-        return new PreservedMaterialGroupingKey(
-            null,
-            normalizedMaterial.BaseColor,
-            normalizedMaterial.MaterialType,
-            normalizedMaterial.TexturePayload,
-            normalizedMaterial.TextureSourceKind,
-            normalizedMaterial.TerrainOverlay,
-            normalizedMaterial.Projection,
-            normalizedMaterial.DepthOffset,
-            normalizedMaterial.TextureScale,
-            normalizedMaterial.TextureOffset,
-            normalizedMaterial.AssetScope,
-            normalizedMaterial.Family,
-            normalizedMaterial.BundledVariantIndex,
-            normalizedMaterial.TerrainMeshCode);
-    }
-
-    private static ResoniteMaterialBinding NormalizePreservedMaterial(ResoniteMaterialBinding material)
-    {
-        return ResoniteSceneMaterialConventions.NormalizeBatchGroupedMaterialBinding(material);
-    }
-
     private async Task EmitAtlasBatchAsync(
         SourceFileBatchKey sourceFileKey,
         IReadOnlyList<CityObjectBakeCandidate> batchCandidates,
@@ -1323,22 +1278,6 @@ internal sealed class NonDemCityObjectBaker(
         string PolicyContext,
         string SourceFileRelativePath);
 
-    private readonly record struct PreservedMaterialGroupingKey(
-        DefaultCommonMaterialMember? CommonMaterial,
-        ResoniteColor BaseColor,
-        ResoniteMaterialType MaterialType,
-        ResoniteTexturePayload? TexturePayload,
-        ResoniteTextureSourceKind TextureSourceKind,
-        TerrainTextureOverlay? TerrainOverlay,
-        ResoniteMaterialProjection Projection,
-        ResoniteMaterialDepthOffset? DepthOffset,
-        ResoniteFloat2? TextureScale,
-        ResoniteFloat2? TextureOffset,
-        ResoniteMaterialAssetScope AssetScope,
-        string? Family,
-        int? BundledVariantIndex,
-        string? TerrainMeshCode);
-
     private sealed class SourceFileBatchKeyComparer : IComparer<SourceFileBatchKey>
     {
         internal static readonly SourceFileBatchKeyComparer Instance = new();
@@ -1379,75 +1318,4 @@ internal sealed class NonDemCityObjectBaker(
         }
     }
 
-    private sealed class PreservedMaterialGroupingKeyComparer :
-        IEqualityComparer<PreservedMaterialGroupingKey>
-    {
-        internal static readonly PreservedMaterialGroupingKeyComparer Instance = new();
-
-        public bool Equals(PreservedMaterialGroupingKey x, PreservedMaterialGroupingKey y)
-        {
-            if (!EqualityComparer<DefaultCommonMaterialMember?>.Default.Equals(x.CommonMaterial, y.CommonMaterial))
-            {
-                return false;
-            }
-
-            if (x.CommonMaterial is not null)
-            {
-                return ReferenceEquals(x.TexturePayload, y.TexturePayload)
-                    && x.TextureSourceKind == y.TextureSourceKind
-                    && EqualityComparer<TerrainTextureOverlay?>.Default.Equals(x.TerrainOverlay, y.TerrainOverlay)
-                    && string.Equals(x.TerrainMeshCode, y.TerrainMeshCode, StringComparison.Ordinal);
-            }
-
-            return x.BaseColor == y.BaseColor
-                && x.MaterialType == y.MaterialType
-                && ReferenceEquals(x.TexturePayload, y.TexturePayload)
-                && x.TextureSourceKind == y.TextureSourceKind
-                && EqualityComparer<TerrainTextureOverlay?>.Default.Equals(x.TerrainOverlay, y.TerrainOverlay)
-                && x.Projection == y.Projection
-                && EqualityComparer<ResoniteMaterialDepthOffset?>.Default.Equals(x.DepthOffset, y.DepthOffset)
-                && EqualityComparer<ResoniteFloat2?>.Default.Equals(x.TextureScale, y.TextureScale)
-                && EqualityComparer<ResoniteFloat2?>.Default.Equals(x.TextureOffset, y.TextureOffset)
-                && x.AssetScope == y.AssetScope
-                && string.Equals(x.Family, y.Family, StringComparison.Ordinal)
-                && x.BundledVariantIndex == y.BundledVariantIndex
-                && string.Equals(x.TerrainMeshCode, y.TerrainMeshCode, StringComparison.Ordinal);
-        }
-
-        public int GetHashCode(PreservedMaterialGroupingKey obj)
-        {
-            HashCode hash = new();
-            hash.Add(obj.CommonMaterial);
-            if (obj.CommonMaterial is not null)
-            {
-                if (obj.TexturePayload is not null)
-                {
-                    hash.Add(RuntimeHelpers.GetHashCode(obj.TexturePayload));
-                }
-                hash.Add(obj.TextureSourceKind);
-                hash.Add(obj.TerrainOverlay);
-                hash.Add(obj.TerrainMeshCode, StringComparer.Ordinal);
-                return hash.ToHashCode();
-            }
-
-            hash.Add(obj.BaseColor);
-            hash.Add(obj.MaterialType);
-            if (obj.TexturePayload is not null)
-            {
-                hash.Add(RuntimeHelpers.GetHashCode(obj.TexturePayload));
-            }
-            hash.Add(obj.TextureSourceKind);
-            hash.Add(obj.TerrainOverlay);
-            hash.Add(obj.Projection);
-            hash.Add(obj.DepthOffset);
-            hash.Add(obj.TextureScale);
-            hash.Add(obj.TextureOffset);
-            hash.Add(obj.AssetScope);
-            hash.Add(obj.Family, StringComparer.Ordinal);
-            hash.Add(obj.BundledVariantIndex);
-            hash.Add(obj.TerrainMeshCode, StringComparer.Ordinal);
-            return hash.ToHashCode();
-        }
-
-    }
 }
